+++ constant/product.go
@@ -1,8 +1,9 @@
 package constant
 
 const (
-	StatusApproved   = "APPROVED"
-	StatusWaiting    = "WAITING LIST"
-	StatusRejected   = "REJECTED"
-	StatusSubscribed = "SUBSCRIBED"
+	StatusApproved       = "APPROVED"
+	StatusWaiting        = "WAITING LIST"
+	StatusRejected       = "REJECTED"
+	StatusWaitingPayment = "PAYMENT REVIEW"
+	StatusSubscribed     = "SUBSCRIBED"
 )